Extract common logic into cgroup base scraper
