+++ datadog_checks_base/datadog_checks/base/checks/cgroup/__init__.py
@@ -1,0 +1,8 @@
+# (C) Datadog, Inc. 2019
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+
+from .scraper import CgroupMetricsScraper
+
+__all__ = ['CgroupMetricsScraper']
